requirements.txt
# 核心依赖
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-openai-like>=0.1.0

# Web框架
flask>=3.0.0
flask-cors>=4.0.0

# 环境变量
python-dotenv>=1.0.0

# 文件处理
pypdf>=3.17.0
python-docx>=1.1.0

# 工具库
werkzeug>=3.0.0

# 可选：向量数据库
# chromadb>=0.4.0
# qdrant-client>=1.7.0
# pinecone-client>=3.0.0

# 可选：增强PDF处理
# pymupdf>=1.23.0
# pdfminer.six>=20221105

# 可选：OCR支持
# pytesseract>=0.3.10
# pdf2image>=1.16.3
